// .storyblok/types/storyblok-components.d.ts
// This file was generated by the storyblok CLI.
// DO NOT MODIFY THIS FILE BY HAND.
export interface Calendar {
  resources?: (Calendar | Event | Header | Page | Resource)[];
  events?: (Calendar | Event | Header | Page | Resource)[];
  component: "calendar";
  _uid: string;
  [k: string]: unknown;
}

export interface Event {
  id?: string;
  name?: string;
  readOnly?: boolean;
  resourceId?: string;
  draggable?: boolean;
  resizable?: boolean;
  allDay?: boolean;
  startDate?: string;
  endDate?: string;
  exceptionDates?: string;
  recurrenceRule?: string;
  cls?: string;
  eventColor?: string;
  eventStyle?: string;
  iconCls?: string;
  style?: string;
  component: "event";
  _uid: string;
  [k: string]: unknown;
}

export interface Header {
  title?: string;
  component: "header";
  _uid: string;
  [k: string]: unknown;
}

export interface Page {
  body?: (Calendar | Event | Header | Page | Resource)[];
  component: "page";
  _uid: string;
  [k: string]: unknown;
}

export interface Resource {
  id?: string;
  name?: string;
  eventColor?: string;
  readOnly?: boolean;
  component: "resource";
  _uid: string;
  [k: string]: unknown;
}